refactor: move IdxOrName to a types submodule
