+++ src/types/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod dtype;
 pub(crate) mod excelreader;
 pub(crate) mod excelsheet;
+pub(crate) mod idx_or_name;
 pub(crate) use excelreader::ExcelReader;
 pub(crate) use excelsheet::ExcelSheet;
